# server/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Loventia API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Loventia API Team
    url: https://loventia.example.com
    email: api@loventia.example.com
  description: |
    OpenAPI specification for the Loventia dating application.

    This spec is aligned with the current server/src/app.js routing we just test-drove
    with PowerShell:

      - Primary auth endpoint: POST /api/auth/login
      - Legacy/fallback auth endpoint: POST /api/users/login
      - Refresh: POST /api/auth/refresh
      - Self: GET /api/auth/me and GET /api/users/me and GET /api/me
      - Reset-password flow: POST /api/auth/forgot-password, POST /api/auth/reset-password
      - Private diagnostics: GET /api/auth/private/ping, GET /api/auth/private/no-me
      - User photo management (the things we just tested): 
          * DELETE /api/users/{userId}/photos/{slot}
          * DELETE /api/users/{userId}/photos?path=...
          * PUT    /api/users/{userId}/photos/reorder
          * POST   /api/users/{userId}/set-avatar
      - Billing endpoints (Stripe webhook at /webhooks/stripe and legacy /api/payment/stripe-webhook)
      - Admin CSV exports
      - Health/readiness endpoints

    IMPORTANT:
      - We **document both** logins on purpose:
          1) /api/auth/login  (main, preferred)
          2) /api/users/login (legacy/fallback)
      - This matches the README change you will add in server/README.md
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SendMessageBody:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          example: "Hey! Fancy a coffee this weekend?"
        clientId:
          type: string
          description: Optional client-generated idempotency key
paths:
...
  /api/messages/overview:
    get:
      tags: [Messages]
      security: [{ bearerAuth: [] }]
      summary: List conversation summaries
      description: Returns overview cards for recent conversations.
      operationId: messages_overview_list
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
  /api/messages/can-send-intro/{userId}:
    get:
      tags: [Messages]
      security: [{ bearerAuth: [] }]
      summary: Whether current user can send intro to userId
      description: Returns gating decision for sending the first message.
      operationId: messages_can_send_intro
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '400': { description: Invalid user id }
        '401': { description: Unauthorized }
  /api/messages/{userId}:
    get:
      tags: [Messages]
      security: [{ bearerAuth: [] }]
      summary: Get a thread with userId
      description: Returns message thread between the current user and userId.
      operationId: messages_get_thread
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '400': { description: Invalid user id }
        '401': { description: Unauthorized }
    post:
      tags: [Messages]
      security: [{ bearerAuth: [] }]
      summary: Send a message to userId
      description: Sends a message to the given userId.
      operationId: messages_send
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageBody'
      responses:
        '201': { description: Created }
        '400': { description: Invalid payload or user id }
        '401': { description: Unauthorized }
...
security:
  - bearerAuth: []